docstore: add option to limit outstanding RPCs for action lists
This does everything buy dynamo, which is forthcoming.

See #2133.
